% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{air}
\alias{air}
\title{Air density measurements}
\format{
Classes data.table and data.frame:	5 observations of 7 variables:
\describe{
 \item{date}{Date, R class "Date", format \code{YYYY-MM-DD}.}
 \item{trial}{Character, label a through e.}
 \item{humidity}{Factor, low, medium, or high.}
 \item{temperature}{Numeric, [K] }
 \item{pressure}{Numeric, atmospheric pressure [Pa] }
 \item{R}{Integer, gas constant
          [J kg\eqn{^{-1}}{^{-1}}K\eqn{^{-1}}{^{-1}}] }
 \item{density}{Numeric, calculated air density
       [kg m\eqn{^{-3}}{^{-13}}] }
}
}
\usage{
data(air, package = "formatdown")
}
\description{
A table of air properties at room temperature, simulating multiple
measurements at approximately steady state,
}
\keyword{datasets}
