% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{air_meas}
\alias{air_meas}
\title{Air density measurements}
\format{
Classes data.table and data.frame:	5 observations of 7 variables:
\describe{
 \item{date}{Date, R class "Date", format \code{YYYY-MM-DD}.}
 \item{trial}{Character, label a through e.}
 \item{humid}{Factor, low, med, or high.}
 \item{temp}{Numeric, [K] }
 \item{pres}{Numeric, atmospheric pressure [Pa] }
 \item{sp_gas}{Numeric, specific gas constant, ideal gas constant in mass form
          [J kg\eqn{^{-1}}{^{-1}}K\eqn{^{-1}}{^{-1}}] }

 \item{dens}{Numeric, calculated air density
       [kg m\eqn{^{-3}}{^{-13}}] }
}
}
\usage{
data(air_meas, package = "formatdown")
}
\description{
A table of air properties at room temperature, simulating multiple
measurements at approximately steady state,
}
\keyword{datasets}
