% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atmosphere}
\alias{atmosphere}
\title{Properties of standard atmosphere}
\format{
Classes data.table and data.frame: 10 observations of 5 variables:
\describe{
 \item{alt}{Numeric, altitude [m]   }
 \item{temp}{Numeric, air temperature [K]   }
 \item{pres}{Numeric, atmospheric pressure [Pa]  }
 \item{dens}{Numeric, air density [kg m\eqn{^{-3}}{^{-13}}] }
 \item{spd_snd}{Numeric, speed of sound [m/s] }
}
}
\usage{
data(atmosphere, package = "formatdown")
}
\description{
A table of atmospheric properties as a function of altitude from
sea level to approximately 14 km.
}
\keyword{datasets}
