% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{density}
\alias{density}
\title{Air density measurements}
\format{
A data frame with 5 rows and 7 columns:
\describe{
 \item{date}{Date, R class "Date", format \code{YYYY-MM-DD}.}
 \item{trial}{Character, label a--e.}
 \item{humidity}{Factor, low, medium, or high.}
 \item{T_K}{Numeric, temperature, units K.}
 \item{p_Pa}{Numeric, atmospheric pressure, units Pa.}
 \item{R}{Integer, gas constant, units
          J kg\eqn{^{-1}}{^{-1}}K\eqn{^{-1}}{^{-1}}.}
 \item{density}{Numeric, calculated air density, units
       kg/m\eqn{^{3}}{^3}.}
}
}
\usage{
data(density, package = "formatdown")
}
\description{
A data.table of air properties at room temperature, simulating multiple
measurements at approximately steady state,
}
\keyword{datasets}
