% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{element_set}
\alias{element_set}
\title{Chemical elements}
\format{
Classes data.table and data.frame with	354
observations of 4 variables:

\describe{
\item{element}{Character, element name, hydrogen (atomic number 1)
through oganesson (atomic number 118).}
\item{symbol}{Character, element symbol.}
\item{atomic_number}{Character, 1-118.}
\item{mass_number}{Character, 1-294.}
}
}
\usage{
data(element_set, package = "formatdown")
}
\description{
Table of 118 chemical elements with name, symbol, atomic number, and
mass numbers. Elements with more than one possible mass number have
separate rows for each isotope.
}
\details{
In "AZE" or "AZX" notation, "A" is the mass number, "Z" is the atomic number,
and "E" or "X" is the element symbol.
}
\keyword{datasets}
