% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_power.R
\name{format_power}
\alias{format_power}
\title{Format powers of ten}
\usage{
format_power(
  x,
  sig_dig = 3,
  ...,
  engr_notation = TRUE,
  disable_exponent = c(0.1, 1000)
)
}
\arguments{
\item{x}{Numerical vector to be formatted.}

\item{sig_dig}{Scalar integer, number of significant digits, default 3.}

\item{...}{Not used, force later arguments to be used by name.}

\item{engr_notation}{Logical, default TRUE. Places decimal points such that
exponents are multiples of three. If FALSE, result has one digit to
the left of the decimal point.}

\item{disable_exponent}{Numeric vector of length two with the lower and
upper limits over which scientific notation is disabled, default
c(0.1, 1000).}
}
\value{
A character vector with the following properties:
\itemize{
    \item Math powers of ten notation substituted for computer notation.
    \item Delimited with \code{$...$} for output as a mathematical
          expression to R Markdown or Quarto Markdown document.
}
}
\description{
Format a numerical vector as a character string with power of ten notation
in mathematical form or engineering form.
}
\details{
In engineering notation, exponents are multiples of three. Compatible with
many of the output formats of R Markdown and Quarto Markdown. Can be applied
to the numerical columns of a data frame via \code{lapply()}.
}
\examples{
# Scalar value
format_power(101100)

# Vector value
x <- c(1.2222e-6, 2.3333e-5, 3.4444e-4, 4.1111e-3, 5.2222e-2, 6.3333e-1,
       7.4444e+0, 8.1111e+1, 9.2222e+2, 1.3333e+3, 2.4444e+4, 3.1111e+5, 4.2222e+6)
format_power(x)

# Compare significant digits
format_power(x[1], 3)
format_power(x[1], 4)

# Compare notation type
format_power(x[3], engr_notation = TRUE)
format_power(x[3], engr_notation = FALSE)

# Compare disabled exponents limits
format_power(x[6], disable_exponent = c(0.1, 1000))
format_power(x[6], disable_exponent = c(1, 1000))

# Apply to columns of a data frame (data.table syntax)
y <- x[1:6]
z <- x[8:13]
DT <- data.table::data.table(y, z)
DT[, lapply(.SD, function(x) format_power(x))]
}
\concept{format_*}
