% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{format_sci}
\alias{format_sci}
\title{Format scientific notation}
\usage{
format_sci(
  x,
  digits = 4,
  ...,
  omit_power = c(-1, 2),
  set_power = NULL,
  size = formatdown_options("size"),
  delim = formatdown_options("delim"),
  decimal_mark = formatdown_options("decimal_mark"),
  big_mark = formatdown_options("big_mark"),
  small_mark = formatdown_options("small_mark")
)
}
\arguments{
\item{x}{Numerical vector to be formatted. Can be a scalar, a vector, or a
column from a data frame.}

\item{digits}{Integer scalar from 1 through 20 that controls the number of
significant digits in printed numeric values; passed to \code{signif()}.
Default is 4.}

\item{...}{Not used for values, forces subsequent arguments to be referable
only by name.}

\item{omit_power}{Numeric vector \code{c(p, q)} with \code{p <= q}, specifying the
range of exponents over which power-of-ten notation is omitted in either
scientific or engineering power-of-ten format. Default is \code{c(-1, 2)}. If a
single value is assigned, i.e., \code{omit_power = p}, the argument is
interpreted as \code{c(p, p)}. If \code{NULL} or \code{NA}, all elements are formatted in
power-of-ten notation. Argument is overridden by a non-empty \code{set_power}.}

\item{set_power}{Integer scalar. Formats all values in \code{x} with the same
power-of-ten exponent. Default NULL. Assigning a value to set_power
overrides \code{omit_power} arguments.}

\item{size, delim, decimal_mark, big_mark, small_mark}{Formatting options. For
details, see the help page for \code{formatdown_options()}.}
}
\value{
A character vector with the following properties:
\itemize{
\item Numbers represented in scientific powers of ten notation.
\item Elements delimited as inline math markup.
}
}
\description{
Convert a numeric vector to a character vector in which the numbers are
formatted in power-of-ten notation in scientific form and delimited for
rendering as inline equations in an R markdown document.
}
\details{
A convenience function to create scientific power-of-ten notation with a
minimal set of arguments by wrapping \code{format_numbers()}. Use
\code{formatdown_options()} to globally set \code{size}, \code{delim}, \code{decimal_mark},
\code{big_mark}, or \code{small_mark} for use with this convenience function.
Otherwise, use \code{format_numbers()} for direct local access to all arguments.

Arguments after the dots (\code{...}) must be referred to by name.
}
\seealso{
Other format_*: 
\code{\link{format_dcml}()},
\code{\link{format_engr}()},
\code{\link{format_numbers}()},
\code{\link{format_text}()},
\code{\link{format_units}()}
}
\concept{format_*}
