% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_decimal.R, R/format_power.R,
%   R/format_units.R, R/formatdown-deprecated.R
\name{format_decimal}
\alias{format_decimal}
\alias{format_power}
\alias{format_units}
\alias{formatdown-deprecated}
\title{Deprecated functions in package formatdown.}
\usage{
format_decimal(x, digits = 4, ..., big_mark = NULL, delim = "$")

format_power(
  x,
  digits = 4,
  ...,
  format = NULL,
  size = NULL,
  omit_power = c(-1, 2),
  set_power = NULL,
  delim = "$"
)

format_units(
  x,
  digits = 1,
  ...,
  unit = NULL,
  unit_form = NULL,
  big_mark = NULL
)
}
\description{
The functions listed below are deprecated. Alternative functions with
similar functionality are mentioned. Help pages for deprecated functions
are available at \code{help("<function>-deprecated")}.
}
\section{\code{format_decimal}}{

For \code{format_decimal()}, use \code{format_dcml()} or
\code{format_numbers(..., format = "dcml")}
}

\section{\code{format_power}}{

For \code{format_power()}, use \code{format_numbers()}, \code{format_sci()} or
\code{format_engr()}.
}

\section{\code{format_units}}{

For \code{format_units()}, use \code{format_numbers()} after first assigning
physical measurement units using the units package.
}

\keyword{internal}
