% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatdown_options.R
\name{formatdown_options}
\alias{formatdown_options}
\title{Options settings}
\usage{
formatdown_options(...)
}
\arguments{
\item{...}{One or more \code{name = value} pairs to set values; or one or more
quoted option names to get values.}
}
\description{
Set and examine several global options which affect the way in which
formatdown displays its results.
}
\details{
Global options are provided for those arguments most likely to be used
throughout a document. For example, a period (".") is the convention for
the decimal separator in many regions, while in others a comma (",") is the
convention. A user would likely prefer to set this option once rather than
restating it in every function call.

Globally-set arguments can be overridden locally by assigning them in a
function call.
}
\section{formatdown options}{

\describe{

\item{\code{size}}{Font size. Possible values are NA (default), "scriptsize",
"small", "normalsize", "large", and "huge". If not NA, adds a prefix to
the markup to invoke the macros \verb{\\scriptsize}, \verb{\\small}, \verb{\\normalsize},
etc.}

\item{\code{delim}}{Character vector of length one or two defining the math markup
delimiters. Possible values include \code{c("$", "$")} or \code{c("\\\\(", "\\\\)")}.
Alternatively, one can use \code{"$"} (default) or \code{"\\\\("}, both of which
create appropriate left and right delimiters. If required by one's markup
environment, custom 2-element (left and right) delimiters can be assigned.}

\item{\code{decimal_mark}}{Character. Possible values are a period (\code{"."},
default) or a comma (\code{","}) to denote the numeric decimal point; passed
to \code{formatC(decimal.mark)}.}

\item{\code{big_mark}}{Character. Possible values are empty (\code{""}, default)
or "thin" to produce a LaTeX-style thin space (\code{"\\\\\\\\,"}) to separate
numbers into groups of three to the left of the decimal mark; passed to
\code{formatC(big.mark)}.}

\item{\code{small_mark}}{c}

\item{\code{word_space}}{Character. Defines the LaTeX-style math-mode macro to
preserve a horizontal space between words. Default is (\code{"\\\\\\\\>"}).
Alternatives include (\code{"\\\\\\\\:"}) or ("\verb{\\\\\\\\ }").}

}
}

\examples{
# get default options
formatdown_options("size", "delim")

# set new options
formatdown_options(size = "large", delim = c("\\\\(", "\\\\)"))

# verify new options
formatdown_options("size", "delim")

# reset to default options
reset_formatdown_options()

# confirm default options
formatdown_options("size", "delim")
}
\seealso{
Other *_options: 
\code{\link{reset_formatdown_options}()}
}
\concept{*_options}
