% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metals}
\alias{metals}
\title{Properties of metals}
\format{
Classes data.table and data.frame:	5 observations of  5 variables:
\describe{
 \item{metal}{Character, name of material}
 \item{dens}{Numeric, density [kg m\eqn{^{-3}}{^{-3}}] }
 \item{thrm_exp}{Numeric, coefficient of thermal expansion [m m\eqn{^{-1}}{^{-1}}K\eqn{^{-1}}{^{-1}}] }
 \item{thrm_cond}{Numeric, thermal conductivity [W m\eqn{^{-1}}{^{-1}}K\eqn{^{-1}}{^{-1}}] }
 \item{elast_mod}{Numeric, modulus of elasticity [Pa] }
}
}
\usage{
data(metals, package = "formatdown")
}
\description{
A table of basic properties of several metals
}
\keyword{datasets}
