% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatdown-package.R
\docType{package}
\name{powernotation-package}
\alias{powernotation-package}
\title{Formatting Tools for Markdown}
\description{
Format a numerical vector as a character string with power of ten notation
in mathematical form or engineering form. In engineering notation, exponents
are multiple of three. Compatible with many of the output formats of R
Markdown and Quarto Markdown. Can be applied to the numerical columns of a
data frame via \code{lapply()}.
}
\concept{package}
