% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatdown_options.R
\name{reset_formatdown_options}
\alias{reset_formatdown_options}
\title{Reset formatdown options}
\usage{
reset_formatdown_options()
}
\description{
Reset formatdown options to their default values.
}
\seealso{
Other *_options: 
\code{\link{formatdown_options}()}
}
\concept{*_options}
