% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water}
\alias{water}
\title{Properties of water}
\format{
Classes data.table and data.frame:	11 observations of 5 variables:
\describe{
 \item{temp}{Numeric, temperature [K] }
 \item{dens}{Numeric, density  [kg m\eqn{^{-3}}{^{-3}}] }
 \item{sp_wt}{Numeric, specific weight [N m\eqn{^{-3}}{^{-3}}] }
 \item{visc}{Numeric, dynamic viscosity  [Pa-s] }
 \item{bulk_mod}{Numeric, bulk modulus  [Pa] }
}
}
\source{
E. Maurer E and I. Embry (2022) \emph{hydraulics: Basic Pipe and
Open Channel Hydraulics}, R package ver. 0.6.0,
\url{https://edm44.github.io/hydraulics/}.
}
\usage{
data(water, package = "formatdown")
}
\description{
Table of water properties at atmospheric pressure as a function of
temperature.
}
\keyword{datasets}
