% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atmosphere}
\alias{atmosphere}
\title{Properties of standard atmosphere}
\format{
Classes data.table and data.frame: 10 observations of 5 variables:
\describe{
 \item{altitude}{Numeric    [m] }
 \item{temperature}{Numeric [K] }
 \item{pressure}{Numeric    [Pa] }
 \item{density}{Numeric     [kg m\eqn{^{-3}}{^{-13}}] }
 \item{speed_sound}{Numeric [m/s] }
}
}
\usage{
data(atmosphere, package = "formatdown")
}
\description{
A table of atmospheric properties as a function of altitude from
sea level to approximately 14 km.
}
\keyword{datasets}
