% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{density}
\alias{density}
\title{Ideal gas sample data frame}
\format{
A data frame with 5 rows and 6 columns:
\describe{
 \item{date}{A date variable.}
 \item{trial}{A character variable.}
 \item{humidity}{A factor variable.}
 \item{T_K}{Numeric air temperature in Kelvin.}
 \item{p_Pa}{Numeric air pressure in Pascals.}
 \item{R}{Numeric gas constant in J/(kg K)}
 \item{density}{Numeric air density in kg/m^3}
}
}
\usage{
data(density, package = "formatdown")
}
\description{
Air density at room temperature. A data frame to illustrate features of
the \code{format_power()}.
}
\keyword{datasets}
