% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{format_dcml}
\alias{format_dcml}
\title{Format decimal notation}
\usage{
format_dcml(
  x,
  digits = 4,
  ...,
  size = formatdown_options("size"),
  delim = formatdown_options("delim"),
  decimal_mark = formatdown_options("decimal_mark"),
  big_mark = formatdown_options("big_mark"),
  small_mark = formatdown_options("small_mark")
)
}
\arguments{
\item{x}{Numerical vector to be formatted. Can be a scalar, a vector, or a
column from a data frame.}

\item{digits}{Integer scalar from 1 through 20 that controls the number of
significant digits in printed numeric values; passed to \code{signif()}.
Default is 4.}

\item{...}{Not used for values, forces subsequent arguments to be referable
only by name.}

\item{size, delim, decimal_mark, big_mark, small_mark}{Formatting options. For
details, see the help page for \code{formatdown_options()}.}
}
\value{
A character vector with the following properties:
\itemize{
\item Numbers represented in decimal notation.
\item Elements delimited as inline math markup.
}
}
\description{
Convert a numeric vector to a character vector in which the numbers are
formatted in decimal form and delimited for rendering as inline equations
in an R markdown document.
}
\details{
A convenience function to create decimal notation with a
minimal set of arguments by wrapping \code{format_numbers()}. Use
\code{formatdown_options()} to globally set \code{size}, \code{delim}, \code{decimal_mark},
\code{big_mark}, or \code{small_mark} for use with this convenience function.
Otherwise, use \code{format_numbers()} for direct local access to all arguments.

Arguments after the dots (\code{...}) must be referred to by name.
}
\seealso{
Other format_*: 
\code{\link{format_engr}()},
\code{\link{format_numbers}()},
\code{\link{format_sci}()},
\code{\link{format_text}()},
\code{\link{format_units}()}
}
\concept{format_*}
