% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatdown-package.R
\docType{package}
\name{formatdown-package}
\alias{formatdown-package}
\alias{formatdown}
\title{Formatting Tools for R Markdown Documents}
\description{
Provides a small set of tools for common formatting tasks when writing
documents in R markdown or Quarto markdown. Works with outputs in html,
pdf, docx and possibly others.
}
\keyword{internal}
