% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water}
\alias{water}
\title{Water properties}
\format{
Classes data.table and data.frame:	5 observations of 5 variables:
\describe{
 \item{temperature}{Numeric, temperature [K] }
 \item{density}{Numeric, density  [kg m\eqn{^{-3}}{^{-3}}] }
 \item{specific_weight}{Numeric, specific weight
       [N m\eqn{^{-3}}{^{-3}}] }
 \item{viscosity}{Numeric, dynamic viscosity  [Pa-s] }
 \item{bulk_modulus}{Numeric, bulk modulus  [Pa] }
}
}
\source{
Excerpt from the R hydraulics package,
        \url{https://CRAN.R-project.org/package=hydraulics}
}
\usage{
data(water, package = "formatdown")
}
\description{
A table with properties of water at different temperatures.
}
\keyword{datasets}
