% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water}
\alias{water}
\title{Water properties}
\format{
A data frame with 5 rows and 5 columns:
\describe{
 \item{temperature}{Numeric, temperature, units K.}
 \item{density}{Numeric, density, units kg/m\eqn{^{3}}{^3}.}
 \item{specific_weight}{Numeric, specific weight, units
       N/m\eqn{^{3}}{^3}.}
 \item{viscosity}{Numeric, dynamic viscosity, units Pa-s.}
 \item{bulk_modulus}{Numeric, bulk modulus, units Pa.}
}
}
\source{
Excerpt from the R hydraulics package,
        \url{https://CRAN.R-project.org/package=hydraulics}
}
\usage{
data(water, package = "formatdown")
}
\description{
A data.table with properties of water at different temperatures.
}
\keyword{datasets}
